refactor(datahub): migrate SecurityStore to domains/metadata/security
迁移 SecurityStore 到新的 Metadata 域结构：

变更：
- 创建 domains/metadata/security/models.py (SecurityRegistration)
- 创建 domains/metadata/security/security_store.py (SecurityStore + _build_in_clause)
- 创建 domains/metadata/security/__init__.py
- 更新 stores/security_store.py 为弃用警告 + 重新导出包装器
- 更新测试中的 logger mock 路径
- 修复 domains 模块的类型检查问题

设计原则：
- 保留所有原有实现不变
- 弃用警告使用 warnings.warn (DeprecationWarning)
- 保持向后兼容性（旧导入路径仍然可用）
- 私有函数 _build_in_clause 不导出
- 使用 TYPE_CHECKING 处理未实现的 MetadataQueryService
- 添加 noqa: E402 允许 warnings.warn 后的导入

验证：
- ✅ 所有单元测试通过 (22 passed)
- ✅ SecurityAccessor 测试通过 (17 passed)
- ✅ 类型检查通过 (0 errors)
- ✅ 代码格式检查通过
