+++ packages/datahub/src/ditto_datahub/domains/metadata/__init__.py
@@ -3,17 +3,15 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ditto_datahub.domains.metadata.metadata_query_service import (
-        MetadataQueryService,
-    )
+    # 类型检查时的导入（将在后续任务实现）
+    MetadataQueryService = object
 else:
     # 运行时导入（将在后续任务实现）
     try:
-        from ditto_datahub.domains.metadata.metadata_query_service import (
+        from ditto_datahub.domains.metadata.metadata_query_service import (  # type: ignore[import-not-found]
             MetadataQueryService,
         )
     except ImportError:
-        # MetadataQueryService 将在后续任务中实现
         MetadataQueryService = None  # type: ignore[misc,assignment]
 
 __all__ = ["MetadataQueryService"]
