+++ packages/datahub/src/ditto_datahub/domains/metadata/security/__init__.py
@@ -1,0 +1,6 @@
+"""Security 子域 - 证券主数据."""
+
+from ditto_datahub.domains.metadata.security.models import SecurityRegistration
+from ditto_datahub.domains.metadata.security.security_store import SecurityStore
+
+__all__ = ["SecurityRegistration", "SecurityStore"]
